refactor(pipeline): move entrypoint from scripts/ to pipeline/entrypoints/
Resolves recurring Docker module import issues — the entrypoint now
lives in the pipeline package and is discoverable via the editable
install. Run as python -m pipeline.entrypoints.generate_dataset.

Fixes #361
Refs #344
